Partial fixup of wall drawing.

+++ draw.h
@@ -12,7 +12,7 @@ void line_move_to(uint16_t x, uint16_t y);
 void line_draw_to(uint8_t color, uint16_t x, uint16_t y);
 
 // Values are actual pixel locations.
-void wall_move_to(uint8_t x, uint8_t y_top, uint8_t y_bot);
-void wall_draw_to(uint8_t color, uint8_t x, uint8_t y_top, uint8_t y_bot);
+void wall_move_to(uint16_t x, uint16_t y_top, uint16_t y_bot);
+void wall_draw_to(uint8_t color, uint16_t x, uint16_t y_top, uint16_t y_bot);
 
 #endif // not DRAW_H
